worker/requirements.txt
openpyxl==3.1.5
num2words==0.5.14
python-dotenv==1.0.1
psycopg[binary]==3.2.12
